[vda-node-manager] Update test script
- Separte `stake_and_lock.test.ts` from the `user_write.test.ts`

[vda-common-test] Added `owner()` function to the `ERC20Manager`
