+++ packages/vda-node-manager/.env.example
@@ -6,7 +6,5 @@ RPC_URL="<Input the RPC_URL of chain that contract deployed>"
 ## Request the contract owner to send Tokens for testing
 PROVIDER_KEY="<Input private key of Provider wallet>"
 
-# Flag to indicate token minting for testing deposit features
-# This can be set as `true` when the above `PRIVATE_KEY` represent the contract owner
-# Others should get received token to the above address of `PROVIDER_KEY`
-NEED_MINT_TOKEN="false"
+# Test chain : One of "POLAMOY", "POLPOS", "DEVNET"
+BLOCKCHAIN_ANCHOR = "POLAMOY"
