+++ packages/vda-node-manager/CHANGELOG.md
@@ -1,10 +1,13 @@
-2024-07-09 (v)
+2024-07-10 (v)
 -------------------
 
 - Add a dependency to the `@verida/vda-client-base`
 - Update test codes for refactoring DIDs
 - Added `Test.md` file
 - Updated .env.example file
+- Separated `stake_and_lock.test.ts` from the `user_write.test.ts`
+- Added `owner()` function 
+- Updated test script to let general users run the `user_write.test.ts`
 
 2024-03-22 (v3.0.2)
 -------------------
