+++ packages/vda-node-manager/test/user_read.test.ts
@@ -2,22 +2,16 @@ require('dotenv').config();
 import { VeridaNodeManager } from "../src/index"
 import { BlockchainAnchor, EnumStatus } from "@verida/types";
 import { BigNumber, Wallet } from 'ethers';
-import { addInitialData, compareNodeData } from "./helpers";
+import { addInitialData, compareNodeData, getRegisteredDataCenterIds } from "./helpers";
 import { REGISTERED_DIDS, DID_NODE_MAP, getBlockchainAPIConfiguration, REGISTERED_DATACENTRES, REMOVED_DATACENTRES, REMOVE_START_DIDS, LOCK_LIST, REGISTERED_LOCK_NODE } from "@verida/vda-common-test";
 
 const assert = require('assert')
 
-// Need to add initial data when the contract first deployed
-const privateKey = process.env.PRIVATE_KEY
-if (!privateKey) {
-    throw new Error('No PRIVATE_KEY in the env file');
-}
-const configuration = getBlockchainAPIConfiguration(privateKey);
-
+const blockchainAnchor = process.env.BLOCKCHAIN_ANCHOR !== undefined ? BlockchainAnchor[process.env.BLOCKCHAIN_ANCHOR] : BlockchainAnchor.POLAMOY;
 
 const createBlockchainAPI = () => {
     return new VeridaNodeManager({
-        blockchainAnchor: BlockchainAnchor.POLAMOY
+        blockchainAnchor
     })
 }
 
@@ -26,8 +20,8 @@ describe('vda-node-manager read only tests', () => {
     let blockchainApi : VeridaNodeManager
 
     before(async () => {
-        registeredCentreIds = await addInitialData(configuration);
         blockchainApi = createBlockchainAPI()
+        registeredCentreIds = await getRegisteredDataCenterIds(blockchainAnchor, blockchainApi);        
     })
 
     describe("Data Centre test", () => {
